@@ -3,17 +3,17 @@ description: A new Flutter package project.
 version: 0.0.1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  bloc: ^8.0.1
-  equatable: ^2.0.3
-  intl: ^0.17.0
-  flutter_bloc: ^8.0.0
+  bloc: ^8.1.0
+  equatable: ^2.0.5
+  intl: ^0.18.0
+  flutter_bloc: ^8.1.1
 
 dev_dependencies:
-  bloc_test: ^9.0.1
+  bloc_test: ^9.1.0
   flutter_test:
     sdk: flutter
