@@ -1,40 +1,26 @@
+// @dart=2.9
+
 import 'package:invalid/invalid.dart';
 
-import 'type_converter.dart';
-import 'validators.dart';
-
-class ValidationConfiguration<
-    DefaultValidationMessagesType extends DefaultValidationMessagesLocalization> {
+class ValidationConfiguration<DefaultValidationMessagesType extends DefaultValidationMessagesLocalization> {
   final List<TypeConverter> _typeConverter;
   final DefaultValidationMessagesType defaultValidationMessagesLocalization;
 
-  static ValidationConfiguration<DefaultValidationMessagesLocalization>
-      _instance;
+  static ValidationConfiguration<DefaultValidationMessagesLocalization> _instance;
 
-  ValidationConfiguration._(
-      this.defaultValidationMessagesLocalization, this._typeConverter);
+  ValidationConfiguration._(this.defaultValidationMessagesLocalization, this._typeConverter);
 
   factory ValidationConfiguration.initialize(
-      {DefaultValidationMessagesType defaultValidationMessages,
-      List<TypeConverter> typeConverter = const []}) {
+      {DefaultValidationMessagesType defaultValidationMessages, List<TypeConverter> typeConverter = const []}) {
     return _instance = ValidationConfiguration<DefaultValidationMessagesType>._(
-        defaultValidationMessages ??
-            EmptyDefaultValidationMessages() as DefaultValidationMessagesType,
-        [
-          StringDoubleTypeConverter(),
-          IntDoubleTypeConverter(),
-          ...typeConverter
-        ]);
+        defaultValidationMessages ?? EmptyDefaultValidationMessages() as DefaultValidationMessagesType,
+        [StringDoubleTypeConverter(), IntDoubleTypeConverter(), ...typeConverter]);
   }
 
-  TypeConverter<dynamic, OutputType> getTypeConverter<OutputType>(
-      Type inputType) {
+  TypeConverter<dynamic, OutputType> getTypeConverter<OutputType>(Type inputType) {
     return _typeConverter.firstWhere(
-            (converter) =>
-                converter.inputType == inputType &&
-                converter.outputType == OutputType,
-            orElse: () => throw UnsupportedError(
-                "Converter from Type $inputType, to $OutputType is missing"))
+            (converter) => converter.inputType == inputType && converter.outputType == OutputType,
+            orElse: () => throw UnsupportedError("Converter from Type $inputType, to $OutputType is missing"))
         as TypeConverter<dynamic, OutputType>;
   }
 
@@ -45,47 +31,34 @@ class ValidationConfiguration<
 }
 
 abstract class DefaultValidationMessagesLocalization {
-  String shouldBeEqualValidationMessage(
-      ShouldBeEqualFormValidator val, Iterable<Field> fields);
+  String shouldBeEqualValidationMessage(ShouldBeEqualFormValidator val, Iterable<Field> fields);
 
-  String shouldBeBetweenOrEqualValidationMessage(
-      ShouldBeBetweenOrEqualValidator val, Field field);
+  String shouldBeBetweenOrEqualValidationMessage(ShouldBeBetweenOrEqualValidator val, Field field);
 
-  String shouldBeBetweenValidationMessage(
-      ShouldBeBetweenValidator val, Field field);
+  String shouldBeBetweenValidationMessage(ShouldBeBetweenValidator val, Field field);
 
-  String shouldBeSmallerOrEqualThanValidationMessage(
-      ShouldBeSmallerOrEqualThenValidator val, Field field);
+  String shouldBeSmallerOrEqualThanValidationMessage(ShouldBeSmallerOrEqualThenValidator val, Field field);
 
-  String shouldBeBiggerOrEqualThanValidationMessage(
-      ShouldBeBiggerOrEqualThenValidator val, Field field);
+  String shouldBeBiggerOrEqualThanValidationMessage(ShouldBeBiggerOrEqualThenValidator val, Field field);
 
-  String shouldBeSmallerThanValidationMessage(
-      ShouldBeSmallerThenValidator val, Field field);
+  String shouldBeSmallerThanValidationMessage(ShouldBeSmallerThenValidator val, Field field);
 
-  String shouldBeBiggerThanValidationMessage(
-      ShouldBeBiggerThanValidator val, Field field);
+  String shouldBeBiggerThanValidationMessage(ShouldBeBiggerThanValidator val, Field field);
 
-  String shouldBeInBetweenDatesValidationMessage(
-      ShouldInBetweenDatesValidator val, Field field);
+  String shouldBeInBetweenDatesValidationMessage(ShouldInBetweenDatesValidator val, Field field);
 
-  String shouldBeFalseValidationMessage(
-      ShouldBeFalseValidator val, Field field);
+  String shouldBeFalseValidationMessage(ShouldBeFalseValidator val, Field field);
 
   String shouldBeTrueValidationMessage(ShouldBeTrueValidator val, Field field);
 
-  String shouldNotBeEmptyOrWhiteSpaceValidationMessage(
-      ShouldNotBeEmptyOrWhiteSpaceValidator val, Field field);
+  String shouldNotBeEmptyOrWhiteSpaceValidationMessage(ShouldNotBeEmptyOrWhiteSpaceValidator val, Field field);
 
-  String shouldNotBeEmptyValidationMessage(
-      ShouldNotBeEmptyValidator val, Field field);
+  String shouldNotBeEmptyValidationMessage(ShouldNotBeEmptyValidator val, Field field);
 
-  String shouldNotBeNullValidationMessage(
-      ShouldNotBeNullValidator val, Field field);
+  String shouldNotBeNullValidationMessage(ShouldNotBeNullValidator val, Field field);
 }
 
-class EmptyDefaultValidationMessages
-    extends DefaultValidationMessagesLocalization {
+class EmptyDefaultValidationMessages extends DefaultValidationMessagesLocalization {
   @override
   String shouldBeEqualValidationMessage(_, __) => "empty val msg";
 
@@ -117,8 +90,7 @@ class EmptyDefaultValidationMessages
   String shouldBeTrueValidationMessage(_, __) => "empty val msg";
 
   @override
-  String shouldNotBeEmptyOrWhiteSpaceValidationMessage(_, __) =>
-      "empty val msg";
+  String shouldNotBeEmptyOrWhiteSpaceValidationMessage(_, __) => "empty val msg";
 
   @override
   String shouldNotBeEmptyValidationMessage(_, __) => "empty val msg";
