@@ -1,7 +1,8 @@
+// @dart=2.9
+
 import 'package:equatable/equatable.dart';
 import 'package:invalid/invalid.dart';
 import 'package:meta/meta.dart';
-import 'validation_configuration.dart';
 
 @immutable
 class ValidationResult<KeyType> extends Equatable {
@@ -32,14 +33,7 @@ class ValidationResult<KeyType> extends Equatable {
   bool get stringify => true;
 
   @override
-  List<Object> get props => [
-        formValidatorKey,
-        validatorKey,
-        fieldKey,
-        isValid,
-        message,
-        isFormValidationResult
-      ];
+  List<Object> get props => [formValidatorKey, validatorKey, fieldKey, isValid, message, isFormValidationResult];
 }
 
 abstract class Validator<KeyType> extends Equatable {
@@ -52,41 +46,31 @@ abstract class Validator<KeyType> extends Equatable {
 }
 
 @immutable
-abstract class FieldValidator<
-    TypeOfValidatedValue,
-    KeyType,
-    TFieldValidator extends FieldValidator<TypeOfValidatedValue, KeyType,
-        TFieldValidator>> extends Validator<KeyType> {
-  final String Function(TFieldValidator validator, Field<KeyType> field)
-      buildErrorMessage;
+abstract class FieldValidator<TypeOfValidatedValue, KeyType,
+    TFieldValidator extends FieldValidator<TypeOfValidatedValue, KeyType, TFieldValidator>> extends Validator<KeyType> {
+  final String Function(TFieldValidator validator, Field<KeyType> field) buildErrorMessage;
 
   bool get allowNull => true;
 
   FieldValidator(this.buildErrorMessage, KeyType key) : super(key);
 
   TypeOfValidatedValue parseValue(dynamic value) =>
-      ValidationConfiguration.instance()
-          .getTypeConverter<TypeOfValidatedValue>(value.runtimeType as Type)
-          .canConvert(value);
+      ValidationConfiguration.instance().getTypeConverter<TypeOfValidatedValue>(value.runtimeType).canConvert(value);
 
   ValidationResult<KeyType> validate(Field<KeyType> field) {
-    if (field.value == null)
-      return createValidationResult(null, allowNull, field);
+    if (field.value == null) return createValidationResult(null, allowNull, field);
     if (field.value.runtimeType == TypeOfValidatedValue)
-      return createValidationResult(field.value as TypeOfValidatedValue,
-          isValid(field.value as TypeOfValidatedValue), field);
+      return createValidationResult(
+          field.value as TypeOfValidatedValue, isValid(field.value as TypeOfValidatedValue), field);
     var parsedValue = parseValue(field.value);
-    return createValidationResult(parsedValue,
-        parsedValue == null ? allowNull : isValid(parsedValue), field);
+    return createValidationResult(parsedValue, parsedValue == null ? allowNull : isValid(parsedValue), field);
   }
 
   @protected
   bool isValid(TypeOfValidatedValue value) => true;
 
-  ValidationResult<KeyType> createValidationResult(
-      TypeOfValidatedValue value, bool isValid, Field<KeyType> field) {
-    return ValidationResult<KeyType>(
-        isValid, buildErrorMessage(this as TFieldValidator, field), false,
+  ValidationResult<KeyType> createValidationResult(TypeOfValidatedValue value, bool isValid, Field<KeyType> field) {
+    return ValidationResult<KeyType>(isValid, buildErrorMessage(this as TFieldValidator, field), false,
         validatorKey: key);
   }
 
@@ -96,62 +80,48 @@ abstract class FieldValidator<
 
 /// A [FormValidator] is a validator, that can span multiple fields.
 @immutable
-abstract class FormValidator<KeyType,
-        TFormValidator extends FormValidator<KeyType, TFormValidator>>
+abstract class FormValidator<KeyType, TFormValidator extends FormValidator<KeyType, TFormValidator>>
     extends Validator<KeyType> {
-  final String Function(
-          TFormValidator validator, Iterable<Field<KeyType>> fields)
-      buildErrorMessage;
+  final String Function(TFormValidator validator, Iterable<Field<KeyType>> fields) buildErrorMessage;
 
   FormValidator(this.buildErrorMessage, {KeyType key}) : super(key);
 
   ValidationResult<KeyType> validate(Iterable<Field<KeyType>> fields) {
-    return ValidationResult<KeyType>(isValid(fields),
-        buildErrorMessage(this as TFormValidator, fields), true,
+    return ValidationResult<KeyType>(isValid(fields), buildErrorMessage(this as TFormValidator, fields), true,
         formValidatorKey: key);
   }
 
   @protected
   bool isValid(Iterable<Field<KeyType>> fields);
-
 }
 
 // Form Validators -----------------------------------------------
-class ShouldBeEqualFormValidator<KeyType>
-    extends FormValidator<KeyType, ShouldBeEqualFormValidator<KeyType>> {
+class ShouldBeEqualFormValidator<KeyType> extends FormValidator<KeyType, ShouldBeEqualFormValidator<KeyType>> {
   final Iterable<KeyType> keysOfFieldsWhichShouldBeEqual;
 
   ShouldBeEqualFormValidator(
       {@required this.keysOfFieldsWhichShouldBeEqual,
-      String Function(ShouldBeEqualFormValidator<KeyType> validator,
-              Iterable<Field> fields)
-          buildErrorMessage,
+      String Function(ShouldBeEqualFormValidator<KeyType> validator, Iterable<Field> fields) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
-                ValidationConfiguration.instance()
-                    .defaultValidationMessagesLocalization
-                    .shouldBeEqualValidationMessage,
+                ValidationConfiguration.instance().defaultValidationMessagesLocalization.shouldBeEqualValidationMessage,
             key: key) {
     assert(keysOfFieldsWhichShouldBeEqual.length >= 2);
   }
 
   @override
   bool isValid(Iterable<Field<KeyType>> fields) {
-    var fieldValues = fields
-        .findByFieldKeys(keysOfFieldsWhichShouldBeEqual)
-        .map<dynamic>((field) => field.value);
+    var fieldValues = fields.findByFieldKeys(keysOfFieldsWhichShouldBeEqual).map<dynamic>((field) => field.value);
 
     return fieldValues.toSet().length <= 1;
   }
 
   @override
   List<Object> get props => [...super.props, keysOfFieldsWhichShouldBeEqual];
-
 }
 
-class MultiFieldDateValidator<KeyType>
-    extends FormValidator<KeyType, MultiFieldDateValidator<KeyType>> {
+class MultiFieldDateValidator<KeyType> extends FormValidator<KeyType, MultiFieldDateValidator<KeyType>> {
   final KeyType dayFieldKey;
   final KeyType monthFieldKey;
   final KeyType yearFieldKey;
@@ -160,14 +130,9 @@ class MultiFieldDateValidator<KeyType>
       {@required this.dayFieldKey,
       @required this.monthFieldKey,
       @required this.yearFieldKey,
-      String Function(MultiFieldDateValidator<KeyType> validator,
-              Iterable<Field> fields)
-          buildErrorMessage,
+      String Function(MultiFieldDateValidator<KeyType> validator, Iterable<Field> fields) buildErrorMessage,
       KeyType key})
-      : super(
-            buildErrorMessage ??
-                (MultiFieldDateValidator<KeyType> _, __) => null,
-            key: key);
+      : super(buildErrorMessage ?? (MultiFieldDateValidator<KeyType> _, __) => null, key: key);
 
   @override
   bool isValid(Iterable<Field<KeyType>> fields) {
@@ -178,38 +143,28 @@ class MultiFieldDateValidator<KeyType>
     var dayInput = dayValue == null ? null : int.tryParse(dayValue);
     var monthInput = monthValue == null ? null : int.tryParse(monthValue);
     var yearInput = yearValue == null ? null : int.tryParse(yearValue);
-    if (monthInput == null || dayInput == null || yearInput == null)
-      return false;
+    if (monthInput == null || dayInput == null || yearInput == null) return false;
     var date = isValidDate(dayInput, monthInput, yearInput);
     return date != null;
   }
 
   DateTime isValidDate(int day, int month, int year) {
     var dateTime = DateTime(year, month, day);
-    return dateTime.year == year &&
-            dateTime.month == month &&
-            dateTime.day == day
-        ? dateTime
-        : null;
+    return dateTime.year == year && dateTime.month == month && dateTime.day == day ? dateTime : null;
   }
 
   @override
   List<Object> get props => [...super.props, dayFieldKey, monthFieldKey, yearFieldKey];
-
 }
 
 // Field Validators ----------------------------------------------
 class ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType>
-    extends FieldValidator<TypeOfValidatedValue, KeyType,
-        ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType>> {
+    extends FieldValidator<TypeOfValidatedValue, KeyType, ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType>> {
   @override
   bool get allowNull => false;
 
   ShouldNotBeNullValidator(
-      {String Function(
-              ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType>
-                  buildErrorMessage,
-              Field field)
+      {String Function(ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType> buildErrorMessage, Field field)
           buildErrorMessage,
       KeyType key})
       : super(
@@ -231,16 +186,11 @@ class ShouldNotBeNullValidator<TypeOfValidatedValue, KeyType>
 
   @override
   List<Object> get props => [...super.props, allowNull];
-
 }
 
-class ShouldNotBeEmptyValidator<KeyType> extends FieldValidator<String, KeyType,
-    ShouldNotBeEmptyValidator<KeyType>> {
+class ShouldNotBeEmptyValidator<KeyType> extends FieldValidator<String, KeyType, ShouldNotBeEmptyValidator<KeyType>> {
   ShouldNotBeEmptyValidator(
-      {String Function(
-              ShouldNotBeEmptyValidator<KeyType> validator, Field field)
-          buildErrorMessage,
-      KeyType key})
+      {String Function(ShouldNotBeEmptyValidator<KeyType> validator, Field field) buildErrorMessage, KeyType key})
       : super(
             buildErrorMessage ??
                 ValidationConfiguration.instance()
@@ -254,12 +204,10 @@ class ShouldNotBeEmptyValidator<KeyType> extends FieldValidator<String, KeyType,
   }
 }
 
-class ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType> extends FieldValidator<
-    String, KeyType, ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType>> {
+class ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType>
+    extends FieldValidator<String, KeyType, ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType>> {
   ShouldNotBeEmptyOrWhiteSpaceValidator(
-      {String Function(ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType> validator,
-              Field field)
-          buildErrorMessage,
+      {String Function(ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -274,17 +222,12 @@ class ShouldNotBeEmptyOrWhiteSpaceValidator<KeyType> extends FieldValidator<
   }
 }
 
-class ShouldBeTrueValidator<KeyType>
-    extends FieldValidator<bool, KeyType, ShouldBeTrueValidator<KeyType>> {
+class ShouldBeTrueValidator<KeyType> extends FieldValidator<bool, KeyType, ShouldBeTrueValidator<KeyType>> {
   ShouldBeTrueValidator(
-      {String Function(ShouldBeTrueValidator<KeyType> validator, Field field)
-          buildErrorMessage,
-      KeyType key})
+      {String Function(ShouldBeTrueValidator<KeyType> validator, Field field) buildErrorMessage, KeyType key})
       : super(
             buildErrorMessage ??
-                ValidationConfiguration.instance()
-                    .defaultValidationMessagesLocalization
-                    .shouldBeTrueValidationMessage,
+                ValidationConfiguration.instance().defaultValidationMessagesLocalization.shouldBeTrueValidationMessage,
             key);
 
   @override
@@ -293,17 +236,12 @@ class ShouldBeTrueValidator<KeyType>
   }
 }
 
-class ShouldBeFalseValidator<KeyType>
-    extends FieldValidator<bool, KeyType, ShouldBeFalseValidator<KeyType>> {
+class ShouldBeFalseValidator<KeyType> extends FieldValidator<bool, KeyType, ShouldBeFalseValidator<KeyType>> {
   ShouldBeFalseValidator(
-      {String Function(ShouldBeFalseValidator<KeyType> validator, Field field)
-          buildErrorMessage,
-      KeyType key})
+      {String Function(ShouldBeFalseValidator<KeyType> validator, Field field) buildErrorMessage, KeyType key})
       : super(
             buildErrorMessage ??
-                ValidationConfiguration.instance()
-                    .defaultValidationMessagesLocalization
-                    .shouldBeFalseValidationMessage,
+                ValidationConfiguration.instance().defaultValidationMessagesLocalization.shouldBeFalseValidationMessage,
             key);
 
   @override
@@ -312,17 +250,15 @@ class ShouldBeFalseValidator<KeyType>
   }
 }
 
-class ShouldInBetweenDatesValidator<KeyType> extends FieldValidator<DateTime,
-    KeyType, ShouldInBetweenDatesValidator<KeyType>> {
+class ShouldInBetweenDatesValidator<KeyType>
+    extends FieldValidator<DateTime, KeyType, ShouldInBetweenDatesValidator<KeyType>> {
   final DateTime max;
   final DateTime min;
 
   ShouldInBetweenDatesValidator(
       {@required this.min,
       @required this.max,
-      String Function(
-              ShouldInBetweenDatesValidator<KeyType> validator, Field field)
-          buildErrorMessage,
+      String Function(ShouldInBetweenDatesValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -338,18 +274,16 @@ class ShouldInBetweenDatesValidator<KeyType> extends FieldValidator<DateTime,
 
   @override
   List<Object> get props => [...super.props, min, max];
-
 }
 
-class ShouldBeBiggerThanValidator<KeyType> extends FieldValidator<double,
-    KeyType, ShouldBeBiggerThanValidator<KeyType>> {
+class ShouldBeBiggerThanValidator<KeyType>
+    extends FieldValidator<double, KeyType, ShouldBeBiggerThanValidator<KeyType>> {
   final double min;
 
   ShouldBeBiggerThanValidator({
     @required this.min,
     KeyType key,
-    String Function(ShouldBeBiggerThanValidator<KeyType> validator, Field field)
-        buildErrorMessage,
+    String Function(ShouldBeBiggerThanValidator<KeyType> validator, Field field) buildErrorMessage,
   }) : super(
             buildErrorMessage ??
                 ValidationConfiguration.instance()
@@ -364,18 +298,15 @@ class ShouldBeBiggerThanValidator<KeyType> extends FieldValidator<double,
 
   @override
   List<Object> get props => [...super.props, min];
-
 }
 
-class ShouldBeSmallerThenValidator<KeyType> extends FieldValidator<double,
-    KeyType, ShouldBeSmallerThenValidator<KeyType>> {
+class ShouldBeSmallerThenValidator<KeyType>
+    extends FieldValidator<double, KeyType, ShouldBeSmallerThenValidator<KeyType>> {
   final double max;
 
   ShouldBeSmallerThenValidator(
       {@required this.max,
-      String Function(
-              ShouldBeSmallerThenValidator<KeyType> validator, Field field)
-          buildErrorMessage,
+      String Function(ShouldBeSmallerThenValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -391,18 +322,15 @@ class ShouldBeSmallerThenValidator<KeyType> extends FieldValidator<double,
 
   @override
   List<Object> get props => [...super.props, max];
-
 }
 
-class ShouldBeBiggerOrEqualThenValidator<KeyType> extends FieldValidator<double,
-    KeyType, ShouldBeBiggerOrEqualThenValidator<KeyType>> {
+class ShouldBeBiggerOrEqualThenValidator<KeyType>
+    extends FieldValidator<double, KeyType, ShouldBeBiggerOrEqualThenValidator<KeyType>> {
   final double min;
 
   ShouldBeBiggerOrEqualThenValidator(
       {@required this.min,
-      String Function(ShouldBeBiggerOrEqualThenValidator<KeyType> validator,
-              Field field)
-          buildErrorMessage,
+      String Function(ShouldBeBiggerOrEqualThenValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -418,18 +346,15 @@ class ShouldBeBiggerOrEqualThenValidator<KeyType> extends FieldValidator<double,
 
   @override
   List<Object> get props => [...super.props, min];
-
 }
 
-class ShouldBeSmallerOrEqualThenValidator<KeyType> extends FieldValidator<
-    double, KeyType, ShouldBeSmallerOrEqualThenValidator<KeyType>> {
+class ShouldBeSmallerOrEqualThenValidator<KeyType>
+    extends FieldValidator<double, KeyType, ShouldBeSmallerOrEqualThenValidator<KeyType>> {
   final double max;
 
   ShouldBeSmallerOrEqualThenValidator(
       {@required this.max,
-      String Function(ShouldBeSmallerOrEqualThenValidator<KeyType> validator,
-              Field field)
-          buildErrorMessage,
+      String Function(ShouldBeSmallerOrEqualThenValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -445,19 +370,16 @@ class ShouldBeSmallerOrEqualThenValidator<KeyType> extends FieldValidator<
 
   @override
   List<Object> get props => [...super.props, max];
-
 }
 
-class ShouldBeBetweenValidator<KeyType>
-    extends FieldValidator<double, KeyType, ShouldBeBetweenValidator<KeyType>> {
+class ShouldBeBetweenValidator<KeyType> extends FieldValidator<double, KeyType, ShouldBeBetweenValidator<KeyType>> {
   final double max;
   final double min;
 
   ShouldBeBetweenValidator(
       {@required this.min,
       @required this.max,
-      String Function(ShouldBeBetweenValidator<KeyType> validator, Field field)
-          buildErrorMessage,
+      String Function(ShouldBeBetweenValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -473,20 +395,17 @@ class ShouldBeBetweenValidator<KeyType>
 
   @override
   List<Object> get props => [...super.props, min, max];
-
 }
 
-class ShouldBeBetweenOrEqualValidator<KeyType> extends FieldValidator<double,
-    KeyType, ShouldBeBetweenOrEqualValidator<KeyType>> {
+class ShouldBeBetweenOrEqualValidator<KeyType>
+    extends FieldValidator<double, KeyType, ShouldBeBetweenOrEqualValidator<KeyType>> {
   final double max;
   final double min;
 
   ShouldBeBetweenOrEqualValidator(
       {@required this.min,
       @required this.max,
-      String Function(
-              ShouldBeBetweenOrEqualValidator<KeyType> validator, Field field)
-          buildErrorMessage,
+      String Function(ShouldBeBetweenOrEqualValidator<KeyType> validator, Field field) buildErrorMessage,
       KeyType key})
       : super(
             buildErrorMessage ??
@@ -502,5 +421,4 @@ class ShouldBeBetweenOrEqualValidator<KeyType> extends FieldValidator<double,
 
   @override
   List<Object> get props => [...super.props, min, max];
-
 }
