@@ -1,3 +1,5 @@
+// @dart=2.9
+
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:invalid/invalid.dart';
@@ -9,8 +11,8 @@ class ValidationResults<FormKeyType> extends StatelessWidget {
   final ValidityFilter filterByValidity;
   final bool ignoreIfFormIsEnabled;
   final bool onlyFirstValidationResult;
-  final Widget Function(List<ValidationResult> validationResults)
-      validationResultsBuilder;
+  final Widget Function(List<ValidationResult> validationResults) validationResultsBuilder;
+
   /// Placeholder widget when the validation results are empty. Defaults to empty [Container]
   final Widget validationResultPlaceholder;
 
@@ -23,8 +25,7 @@ class ValidationResults<FormKeyType> extends StatelessWidget {
       ValidityFilter filterByValidity,
       bool ignoreIfFormIsEnabled,
       bool onlyFirstValidationResult,
-      this.validationResultPlaceholder
-      })
+      this.validationResultPlaceholder})
       : ignoreIfFormIsEnabled = ignoreIfFormIsEnabled ?? false,
         filterByValidity = filterByValidity ?? ValidityFilter.OnlyInvalid,
         padding = padding ?? EdgeInsets.zero,
@@ -33,11 +34,9 @@ class ValidationResults<FormKeyType> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<FormValidationCubit<FormKeyType>,
-        FormValidationState<FormKeyType>>(builder: (context, state) {
-      var filteredValidationResults = ignoreIfFormIsEnabled
-          ? state.allValidationResults
-          : state.validationResultsWhenFormIsEnabled;
+    return BlocBuilder<FormValidationCubit<FormKeyType>, FormValidationState<FormKeyType>>(builder: (context, state) {
+      var filteredValidationResults =
+          ignoreIfFormIsEnabled ? state.allValidationResults : state.validationResultsWhenFormIsEnabled;
 
       switch (filterByValidity) {
         case ValidityFilter.OnlyValid:
@@ -49,23 +48,18 @@ class ValidationResults<FormKeyType> extends StatelessWidget {
         default:
       }
 
-      if (filterByKeys != null)
-        filteredValidationResults =
-            filteredValidationResults.filterByKeys(filterByKeys);
+      if (filterByKeys != null) filteredValidationResults = filteredValidationResults.filterByKeys(filterByKeys);
 
       switch (filterByValidatorType) {
         case ValidatorTypeFilter.FieldValidator:
-          filteredValidationResults =
-              filteredValidationResults.onlyFieldValidationResults;
+          filteredValidationResults = filteredValidationResults.onlyFieldValidationResults;
           break;
         case ValidatorTypeFilter.FormValidator:
-          filteredValidationResults =
-              filteredValidationResults.onlyFormValidationResults;
+          filteredValidationResults = filteredValidationResults.onlyFormValidationResults;
           break;
       }
 
-      if (onlyFirstValidationResult)
-        filteredValidationResults = filteredValidationResults.take(1);
+      if (onlyFirstValidationResult) filteredValidationResults = filteredValidationResults.take(1);
 
       return filteredValidationResults.isEmpty
           ? validationResultPlaceholder ?? Container()
