@@ -1,5 +1,6 @@
+// @dart=2.9
+
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:invalid/invalid.dart';
 
@@ -39,29 +40,22 @@ void main() {
             children: [
               CustomTextField(
                 key: fieldOneKey,
-                validationCapability: TextValidationCapability<FormKeys>(
-                    validationKey: FormKeys.Key1,
-                    validators: [
-                      ShouldNotBeEmptyValidator(
-                        buildErrorMessage: (validator, field) =>
-                            "should not be empty (Key 1)",
-                      ),
-                      ShouldNotBeEmptyOrWhiteSpaceValidator(
-                        buildErrorMessage: (validator, field) =>
-                            "should not be empty or whitespace (Key 1)",
-                      )
-                    ]),
+                validationCapability: TextValidationCapability<FormKeys>(validationKey: FormKeys.Key1, validators: [
+                  ShouldNotBeEmptyValidator(
+                    buildErrorMessage: (validator, field) => "should not be empty (Key 1)",
+                  ),
+                  ShouldNotBeEmptyOrWhiteSpaceValidator(
+                    buildErrorMessage: (validator, field) => "should not be empty or whitespace (Key 1)",
+                  )
+                ]),
               ),
               CustomTextField(
                 key: fieldTwoKey,
-                validationCapability: TextValidationCapability<FormKeys>(
-                    validationKey: FormKeys.Key2,
-                    validators: [
-                      ShouldNotBeEmptyValidator(
-                        buildErrorMessage: (validator, field) =>
-                            "should not be empty (Key 2)",
-                      )
-                    ]),
+                validationCapability: TextValidationCapability<FormKeys>(validationKey: FormKeys.Key2, validators: [
+                  ShouldNotBeEmptyValidator(
+                    buildErrorMessage: (validator, field) => "should not be empty (Key 2)",
+                  )
+                ]),
               ),
               if (validationMessages != null) validationMessages
             ],
@@ -73,31 +67,25 @@ void main() {
   }
 
   group('ValidationForm', () {
-    testWidgets('onFormValidationCubitCreated should provide the current',
-        (tester) async {
+    testWidgets('onFormValidationCubitCreated should provide the current', (tester) async {
       await setUpValidationForm(tester,
-          onFormValidationCubitCreated:
-              expectAsync1<dynamic, FormValidationCubit<FormKeys>>(
+          onFormValidationCubitCreated: expectAsync1<dynamic, FormValidationCubit<FormKeys>>(
             (formValidationCubit) => expect(formValidationCubit, isNotNull),
           ));
     }, timeout: Timeout(Duration(seconds: 1)));
 
-    testWidgets('onFormTurnedValid should be called once the form turns valid',
-        (tester) async {
-      await setUpValidationForm(tester, onFormTurnedValid:
-          expectAsync1<dynamic, FormValidationState<FormKeys>>((state) {
+    testWidgets('onFormTurnedValid should be called once the form turns valid', (tester) async {
+      await setUpValidationForm(tester,
+          onFormTurnedValid: expectAsync1<dynamic, FormValidationState<FormKeys>>((state) {
         expect(state.isValid, true);
       }));
       await tester.enterText(fieldOneFinder, "test");
       await tester.enterText(fieldTwoFinder, "test");
     }, timeout: Timeout(Duration(seconds: 1)));
 
-    testWidgets(
-        'onFormTurnedInValid should be called once the form turns invalid',
-        (tester) async {
+    testWidgets('onFormTurnedInValid should be called once the form turns invalid', (tester) async {
       await setUpValidationForm(tester,
-          onFormTurnedInValid:
-              expectAsync1<dynamic, FormValidationState<FormKeys>>((state) {
+          onFormTurnedInValid: expectAsync1<dynamic, FormValidationState<FormKeys>>((state) {
             expect(state.isValid, false);
           }, count: 2));
 
@@ -110,8 +98,7 @@ void main() {
       await tester.pumpAndSettle();
     }, timeout: Timeout(Duration(seconds: 1)));
 
-    testWidgets('validationMessages should be hidden when form is disabled',
-        (tester) async {
+    testWidgets('validationMessages should be hidden when form is disabled', (tester) async {
       await setUpValidationForm(tester,
           enabled: false,
           validationMessages: CustomValidationMessages(
@@ -120,8 +107,7 @@ void main() {
       expect(find.text("should not be empty (Key 1)"), findsNothing);
     }, timeout: Timeout(Duration(seconds: 1)));
 
-    testWidgets('validationMessages should shown when the form is enabled',
-        (tester) async {
+    testWidgets('validationMessages should shown when the form is enabled', (tester) async {
       await setUpValidationForm(tester,
           enabled: true,
           validationMessages: CustomValidationMessages(
@@ -132,12 +118,8 @@ void main() {
   });
 
   group('ValidationMessages', () {
-    testWidgets(
-        'should show validation messages of all invalid fields and invalid form validators',
-        (tester) async {
-      await setUpValidationForm(tester,
-          enabled: true,
-          validationMessages: CustomValidationMessages<FormKeys>());
+    testWidgets('should show validation messages of all invalid fields and invalid form validators', (tester) async {
+      await setUpValidationForm(tester, enabled: true, validationMessages: CustomValidationMessages<FormKeys>());
 
       // Make one field invalid
       await tester.enterText(fieldOneFinder, "test");
@@ -181,8 +163,7 @@ void main() {
       expect(find.text("should not be empty (Key 2)"), findsNothing);
     });
 
-    testWidgets(
-        'filterByKeys should show only validation messages with keys, that are in the given list',
+    testWidgets('filterByKeys should show only validation messages with keys, that are in the given list',
         (tester) async {
       await setUpValidationForm(tester,
           enabled: true,
@@ -207,8 +188,7 @@ void main() {
       await tester.enterText(fieldOneFinder, "test");
       await tester.pumpAndSettle();
 
-      expect(find.text("should not be empty (Key 1)"),
-          findsNothing); // Field 1 is valid
+      expect(find.text("should not be empty (Key 1)"), findsNothing); // Field 1 is valid
       expect(find.text("should not be empty (Key 2)"), findsOneWidget);
       expect(find.text("should be equal"), findsNothing);
     });
@@ -244,60 +224,56 @@ void main() {
       expect(find.text("should not be empty (Key 2)"), findsOneWidget);
     });
 
-    testWidgets(
-        'when showOnlyTopMessage is true, only the top validation messages should be shown',
-        (tester) async {
+    testWidgets('when showOnlyTopMessage is true, only the top validation messages should be shown', (tester) async {
       await setUpValidationForm(tester,
           enabled: true,
-          validationMessages: CustomValidationMessages<FormKeys>(
-              filterByKeys: [FormKeys.Key1], onlyFirstValidationResult: true));
+          validationMessages:
+              CustomValidationMessages<FormKeys>(filterByKeys: [FormKeys.Key1], onlyFirstValidationResult: true));
 
       expect(find.text("should not be empty (Key 1)"), findsOneWidget);
-      expect(
-          find.text("should not be empty or whitespace (Key 1)"), findsNothing);
+      expect(find.text("should not be empty or whitespace (Key 1)"), findsNothing);
 
       // Make top validator valid
       await tester.enterText(fieldOneFinder, "   ");
       await tester.pumpAndSettle();
 
       expect(find.text("should not be empty (Key 1)"), findsNothing);
-      expect(find.text("should not be empty or whitespace (Key 1)"),
-          findsOneWidget);
+      expect(find.text("should not be empty or whitespace (Key 1)"), findsOneWidget);
     });
   });
 
-  testWidgets(
-      'should show placeholder widget when no validation messages are present',
-          (tester) async {
-        final validationResultsPlaceholderKey = ValueKey("validationResultsPlaceholderKey");
+  testWidgets('should show placeholder widget when no validation messages are present', (tester) async {
+    final validationResultsPlaceholderKey = ValueKey("validationResultsPlaceholderKey");
 
-        await setUpValidationForm(tester,
-            enabled: true,
-            validationMessages: CustomValidationMessages<FormKeys>(
-              validationResultPlaceholder: Container(key: validationResultsPlaceholderKey,),
-            ));
+    await setUpValidationForm(tester,
+        enabled: true,
+        validationMessages: CustomValidationMessages<FormKeys>(
+          validationResultPlaceholder: Container(
+            key: validationResultsPlaceholderKey,
+          ),
+        ));
 
-        // Make fields valid
-        await tester.enterText(fieldOneFinder, "test");
-        await tester.enterText(fieldTwoFinder, "test");
-        await tester.pumpAndSettle();
+    // Make fields valid
+    await tester.enterText(fieldOneFinder, "test");
+    await tester.enterText(fieldTwoFinder, "test");
+    await tester.pumpAndSettle();
 
-        expect(find.byKey(validationResultsPlaceholderKey), findsOneWidget);
-      });
+    expect(find.byKey(validationResultsPlaceholderKey), findsOneWidget);
+  });
 
-  testWidgets(
-      'should show placeholder widget when no validation took place',
-          (tester) async {
-        final validationResultsPlaceholderKey = ValueKey("validationResultsPlaceholderKey");
+  testWidgets('should show placeholder widget when no validation took place', (tester) async {
+    final validationResultsPlaceholderKey = ValueKey("validationResultsPlaceholderKey");
 
-        await setUpValidationForm(tester,
-            enabled: false,
-            validationMessages: CustomValidationMessages<FormKeys>(
-              validationResultPlaceholder: Container(key: validationResultsPlaceholderKey,),
-            ));
+    await setUpValidationForm(tester,
+        enabled: false,
+        validationMessages: CustomValidationMessages<FormKeys>(
+          validationResultPlaceholder: Container(
+            key: validationResultsPlaceholderKey,
+          ),
+        ));
 
-        expect(find.byKey(validationResultsPlaceholderKey), findsOneWidget);
-      });
+    expect(find.byKey(validationResultsPlaceholderKey), findsOneWidget);
+  });
 }
 
 enum FormKeys { Key1, Key2 }
@@ -315,8 +291,7 @@ class _CustomTextFieldState extends State<CustomTextField> {
   @override
   void initState() {
     super.initState();
-    widget.validationCapability
-        .init(context, controller: _textEditingController);
+    widget.validationCapability.init(context, controller: _textEditingController);
   }
 
   @override
@@ -345,12 +320,10 @@ class CustomValidationMessages<FormKeyType> extends StatefulWidget {
       this.onlyFirstValidationResult = false})
       : super(key: key);
   @override
-  CustomValidationMessagesState createState() =>
-      CustomValidationMessagesState<FormKeyType>();
+  CustomValidationMessagesState createState() => CustomValidationMessagesState<FormKeyType>();
 }
 
-class CustomValidationMessagesState<FormKeyType>
-    extends State<CustomValidationMessages> {
+class CustomValidationMessagesState<FormKeyType> extends State<CustomValidationMessages> {
   @override
   Widget build(BuildContext context) {
     return ValidationResults<FormKeyType>(
@@ -362,10 +335,7 @@ class CustomValidationMessagesState<FormKeyType>
       validationResultPlaceholder: widget.validationResultPlaceholder,
       validationResultsBuilder: (validationMessages) {
         return Column(
-          children: [
-            for (ValidationResult result in validationMessages)
-              Text(result.message)
-          ],
+          children: [for (ValidationResult result in validationMessages) Text(result.message)],
         );
       },
     );
